Move utils to common space
