+++ clinicadl/clinicadl/image_level/resume_CNN.py
@@ -6,9 +6,9 @@ from time import time
 import torch
 from torch.utils.data import DataLoader
 
-from .utils import train
 from ..tools.deep_learning.data import MRIDataset, MinMaxNormalization, load_data
 from ..tools.deep_learning import create_model, load_model, load_optimizer, read_json
+from ..tools.deep_learning.cnn_utils import train
 
 parser = argparse.ArgumentParser(description="Argparser for Pytorch 3D CNN")
